Found workaround for g++ 4.1.2 bug - Netsteria now compiles and runs on Ubuntu Feisty.
git-svn-id: svn://svn.pdos.csail.mit.edu/uia/trunk/uia/sst@2589 eeda1704-1414-0410-8491-b7e82234b991

+++ lib/regcli.h
@@ -22,7 +22,7 @@ class RegClient : public QObject
 
 private:
 	// Max time before rereg - 1 hr
-	static const qint64 maxRereg = 60*60*1000000;
+	static const qint64 maxRereg = (qint64)60*60*1000000;
 
 
 	Host *const h;		// Pointer to our per-host state
